Rename `replicon_core` into `core`
It's shorter and it's obvious that it's for replicon.
